refactor(cli): extract cargo_front_door module from main.rs (#339)
Lift `run_cargo_front_door` and its associated cargo arg helpers,
profile-debug-default detection, linker injection, low-disk warning,
and cargo workspace metadata helpers into a dedicated module. Wire
rust_plan, gc, and the test module to import from the new location.

Refs #339
